buffed plot armour, released v1.67

+++ src/gameHelpers.cpp
@@ -72,7 +72,7 @@ void GameHelper::SpawnEnemies(Game& game) noexcept
 
 	GameHelper::GenerateModifierFunctions[GetRandomValue(0, 2)](spawn_count, game.GlobalData->Level, modifiers);
 
-	float level_scale = 1 + game.GlobalData->Level / 10.0f;
+	float level_scale = 1 + static_cast<float>(game.GlobalData->Level) / 10.0f;
 
 	for (size_t i = 0; i < spawn_count; i++)
 		game.Enemies.emplace_back(locations[i].x, locations[i].y, level_scale, game.GlobalData->CurrentLayer, *game.Assets, types[i], modifiers[i]);
